# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "realestate",
        "title": "realestate",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/v1/user/profile": {
            "get": {
                "description": "GetProfile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "GetProfile",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/profileResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "PutProfile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "PutProfile",
                "parameters": [
                    {
                        "description": "profile update",
                        "name": "profile",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/profileRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "$ref": "#/definitions/EmptyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/v1/user/properties": {
            "get": {
                "description": "List properties auth user id.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "properties"
                ],
                "summary": "List properties by auth user id.",
                "parameters": [
                    {
                        "description": "property list body",
                        "name": "Search",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/propertyListRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/DefaultResponseContainer"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/propertyResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create propertyCreateRequest",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "properties"
                ],
                "summary": "Create propertyCreateRequest",
                "parameters": [
                    {
                        "description": "property data",
                        "name": "propertyCreate",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/propertyCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/DefaultResponseContainer"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/propertyResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/user/properties/{id}": {
            "get": {
                "description": "GetUserProp user propertyGetRequest by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "properties"
                ],
                "summary": "GetUserProp user propertyGetRequest by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "propertyGetRequest ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/DefaultResponseContainer"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/propertyResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update PropertyUpdateRequest by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "properties"
                ],
                "summary": "Update PropertyUpdateRequest",
                "parameters": [
                    {
                        "description": "property data",
                        "name": "updateProperty",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/propertyCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/DefaultResponseContainer"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/propertyResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete Property and images",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "properties"
                ],
                "summary": "Delete Property",
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "$ref": "#/definitions/EmptyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "DefaultResponseContainer": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "data": {},
                "totalCount": {
                    "type": "integer"
                }
            }
        },
        "EmptyResponse": {
            "type": "object"
        },
        "ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "address": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "houseNumber": {
                    "type": "string"
                },
                "neighborhood": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "street": {
                    "type": "string"
                },
                "zipCode": {
                    "type": "string"
                }
            }
        },
...
        "profileRequest": {
            "type": "object",
            "properties": {
                "additionalPhones": {
                    "type": "array",
                    "maxItems": 30,
                    "minItems": 2,
                    "items": {
                        "type": "string"
                    }
                },
                "city": {
                    "type": "string",
                    "maxLength": 30,
                    "minLength": 2
                },
                "companyName": {
                    "type": "string",
                    "maxLength": 200,
                    "minLength": 2
                },
                "country": {
                    "type": "string",
                    "maxLength": 30,
                    "minLength": 2
                },
                "latitude": {
                    "type": "string",
                    "maxLength": 15,
                    "minLength": 2
                },
                "longitude": {
                    "type": "string",
                    "maxLength": 15,
                    "minLength": 2
                },
                "name": {
                    "type": "string",
                    "maxLength": 30,
                    "minLength": 2
                },
                "phone": {
                    "type": "string",
                    "maxLength": 30,
                    "minLength": 2
                },
                "photoUrl": {
                    "type": "string",
                    "maxLength": 200,
                    "minLength": 2
                },
                "screenName": {
                    "type": "string",
                    "maxLength": 30,
                    "minLength": 2
                },
                "surname": {
                    "type": "string",
                    "maxLength": 30,
                    "minLength": 2
                },
                "zip": {
                    "type": "string",
                    "maxLength": 15,
                    "minLength": 2
                }
            }
        },
        "profileResponse": {
            "type": "object",
            "properties": {
                "accountId": {
                    "type": "integer"
                },
                "additionalPhones": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "city": {
                    "type": "string"
                },
                "companyName": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "isAccountDeactivated": {
                    "type": "boolean"
                },
                "isAccountLocked": {
                    "type": "boolean"
                },
                "isAppleLinked": {
                    "type": "boolean"
                },
                "isCreatedPassword": {
                    "type": "boolean"
                },
                "isFacebookLinked": {
                    "type": "boolean"
                },
                "isGoogleLinked": {
                    "type": "boolean"
                },
                "isVerifiedEmail": {
                    "type": "boolean"
                },
                "isVerifiedPhone": {
                    "type": "boolean"
                },
                "latitude": {
                    "type": "string"
                },
                "longitude": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "photoUrl": {
                    "type": "string"
                },
                "screenName": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                },
                "zip": {
                    "type": "string"
                }
            }
        },
        "propertyCreateRequest": {
            "type": "object",
            "required": [
                "brokerName",
                "condition",
                "description",
                "fullAddress",
                "homeType",
                "location",
                "price",
                "priceCurrency",
                "propertyType"
            ],
            "properties": {
                "address": {
                    "$ref": "#/definitions/address"
                },
                "appliance": {
                    "type": "boolean"
                },
                "bathroom": {
                    "type": "integer",
                    "maximum": 50,
                    "minimum": 0
                },
                "bedroom": {
                    "type": "integer",
                    "maximum": 50,
                    "minimum": 0
                },
                "brokerName": {
                    "type": "string",
                    "maxLength": 120,
                    "minLength": 0
                },
                "condition": {
                    "type": "integer",
                    "maximum": 20,
                    "minimum": 0
                },
                "description": {
                    "type": "string",
                    "maxLength": 10000,
                    "minLength": 0
                },
                "floor": {
                    "type": "integer",
                    "maximum": 999,
                    "minimum": 0
                },
                "fullAddress": {
                    "type": "string",
                    "maxLength": 1000,
                    "minLength": 0
                },
                "has3DTour": {
                    "type": "boolean"
                },
                "hasAC": {
                    "type": "boolean"
                },
                "hasGarage": {
                    "type": "boolean"
                },
                "hasImages": {
                    "type": "boolean"
                },
                "hasVideo": {
                    "type": "boolean"
                },
                "heating": {
                    "type": "integer",
                    "maximum": 3,
                    "minimum": 0
                },
                "homeSize": {
                    "type": "number",
                    "maximum": 9999,
                    "minimum": 0
                },
                "homeType": {
                    "type": "integer",
                    "maximum": 20,
                    "minimum": 0
                },
                "isActiveStatus": {
                    "type": "boolean"
                },
                "livingSize": {
                    "type": "number",
                    "maximum": 9999,
                    "minimum": 0
                },
                "location": {
                    "$ref": "#/definitions/propertyLocation"
                },
                "lotSize": {
                    "type": "number",
                    "maximum": 9999,
                    "minimum": 0
                },
                "petsAllowed": {
                    "type": "boolean"
                },
                "price": {
                    "type": "number",
                    "maximum": 99999999,
                    "minimum": 0
                },
                "priceCurrency": {
                    "type": "string",
                    "maxLength": 3,
                    "minLength": 0
                },
                "propertyType": {
                    "type": "integer",
                    "maximum": 20,
                    "minimum": 0
                },
                "totalFloors": {
                    "type": "integer",
                    "maximum": 999,
                    "minimum": 0
                },
                "totalParking": {
                    "type": "integer",
                    "maximum": 100,
                    "minimum": 0
                },
                "yearBuild": {
                    "type": "integer",
                    "maximum": 9999,
                    "minimum": 0
                }
            }
        },
        "propertyListRequest": {
            "type": "object",
            "required": [
                "ids"
            ],
            "properties": {
                "ids": {
                    "type": "array",
                    "maxItems": 10,
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "propertyResponse": {
            "type": "object",
            "properties": {
                "actualDays": {
                    "type": "integer"
                },
                "address": {
                    "$ref": "#/definitions/address"
                },
                "appliance": {
                    "type": "boolean"
                },
                "bathroom": {
                    "type": "integer"
                },
                "bedroom": {
                    "type": "integer"
                },
                "brokerName": {
                    "type": "string"
                },
                "condition": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "floor": {
                    "type": "integer"
                },
                "fullAddress": {
                    "type": "string"
                },
                "has3DTour": {
                    "type": "boolean"
                },
                "hasAC": {
                    "type": "boolean"
                },
                "hasGarage": {
                    "type": "boolean"
                },
                "hasImages": {
                    "type": "boolean"
                },
                "hasVideo": {
                    "type": "boolean"
                },
                "heating": {
                    "type": "integer"
                },
                "homeSize": {
                    "type": "number"
                },
                "homeType": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "isActiveStatus": {
                    "type": "boolean"
                },
                "isSavedHome": {
                    "type": "boolean"
                },
                "location": {
                    "$ref": "#/definitions/propertyLocationHome"
                },
                "lotSize": {
                    "type": "number"
                },
                "petsAllowed": {
                    "type": "boolean"
                },
                "price": {
                    "type": "number"
                },
                "priceCurrency": {
                    "type": "string"
                },
                "propertyType": {
                    "type": "integer"
                },
                "totalFloors": {
                    "type": "integer"
                },
                "totalParking": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "yearBuild": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}